+++ src/features/tips/index.ts
@@ -1,0 +1,1 @@
+export { Tips as Tip } from "./tips.component"