+++ src/features/tips/tips.test.js
@@ -1,8 +1,8 @@
 import { render, unmountComponentAtNode } from "react-dom"
 import { act } from "react-dom/test-utils"
-import { Minutes } from "./Minutes"
+import { Tips } from "./tips"
 
-describe("Minutes", () => {
+describe("Tips", () => {
   let container = null
 
   beforeEach(() => {
@@ -21,63 +21,63 @@ describe("Minutes", () => {
   describe("when timer is started", () => {
     it("should display stage for less than 0 minute, but less than 1", () => {
       act(() => {
-        render(<Minutes seconds={1} />, container)
+        render(<Tips seconds={1} />, container)
       })
       expect(container.textContent).toBe("0m")
     })
 
     it("should display stage for less than 1 minute, but less than 2", () => {
       act(() => {
-        render(<Minutes seconds={61} />, container)
+        render(<Tips seconds={61} />, container)
       })
       expect(container.textContent).toBe("1m")
     })
 
     it("should display stage for more than 2 minutes, but less than 3", () => {
       act(() => {
-        render(<Minutes seconds={121} />, container)
+        render(<Tips seconds={121} />, container)
       })
       expect(container.textContent).toBe("2m")
     })
 
     it("should display stage for less than 3 minutes, but less than 5", () => {
       act(() => {
-        render(<Minutes seconds={181} />, container)
+        render(<Tips seconds={181} />, container)
       })
       expect(container.textContent).toBe("3m")
     })
 
     it("should display stage for less than 5 minutes, but less than 8", () => {
       act(() => {
-        render(<Minutes seconds={301} />, container)
+        render(<Tips seconds={301} />, container)
       })
       expect(container.textContent).toBe("5m")
     })
 
     it("should display stage for less than 8 minutes, but less than 13", () => {
       act(() => {
-        render(<Minutes seconds={481} />, container)
+        render(<Tips seconds={481} />, container)
       })
       expect(container.textContent).toBe("8m")
     })
 
     it("should display stage for less than 13 minutes, but less than 21", () => {
       act(() => {
-        render(<Minutes seconds={781} />, container)
+        render(<Tips seconds={781} />, container)
       })
       expect(container.textContent).toBe("13m")
     })
 
     it("should display stage for end of 13th minute (penultimate timer)", () => {
       act(() => {
-        render(<Minutes seconds={780} />, container)
+        render(<Tips seconds={780} />, container)
       })
       expect(container.textContent).toBe("13m")
     })
 
     it("should display stage for end of 21th minute (last timer)", () => {
       act(() => {
-        render(<Minutes seconds={1260} />, container)
+        render(<Tips seconds={1260} />, container)
       })
       expect(container.textContent).toBe("21m")
     })
@@ -86,7 +86,7 @@ describe("Minutes", () => {
   describe("for bugs", () => {
     test("should display properly stage on 11th minute", () => {
       act(() => {
-        render(<Minutes seconds={659} />, container)
+        render(<Tips seconds={659} />, container)
       })
 
       expect(container.textContent).toBe("8m")
@@ -94,7 +94,7 @@ describe("Minutes", () => {
 
     test("should display 8th stage on 12th minute", () => {
       act(() => {
-        render(<Minutes seconds={661} />, container)
+        render(<Tips seconds={661} />, container)
       })
 
       expect(container.textContent).toBe("8m")