+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+### [0.3.7](https://github.com/m0rtyn/pokoy/compare/v0.3.6...v0.3.7) (2021-11-27)
+
+
+### Bug Fixes
+
+* add color state for started timer ([23dd58e](https://github.com/m0rtyn/pokoy/commit/23dd58e0d700c625c2dc3e0848f1c40f69870935))
+* rework and unite components with bottom labels ([4f2c638](https://github.com/m0rtyn/pokoy/commit/4f2c6383f200b2fc399cd150a7b363aef479ea19))
+
 ### [0.3.6](https://github.com/m0rtyn/pokoy/compare/v0.3.1...v0.3.6) (2021-11-27)
 
 ### [0.3.5](https://github.com/m0rtyn/pokoy/compare/v0.3.1...v0.3.5) (2021-11-25)